\name{Connect}
\alias{Connect}
\title{Create connection to exchange}
\description{
  This function responsible for creation connections.
}
\usage{
Connect(host='localhost',port='',key='',sign='')
}
\arguments{
  \item{host}{localhost for terminal connections or name of exchange for web exchange connections}
 }
\keyword{Connect}