\name{getBalance}
\alias{getBalance}
\title{Download account balance}
\description{
  This function responsible for downloading account balance
}
\usage{
getBalance(Connection)
}
\arguments{
  \item{Connection}{localhost for terminal connections and name of exchange for web exchange connections}
 }
\keyword{getBalance}