% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEarnings.R
\name{getEarnings}
\alias{getEarnings}
\title{Download earnings data from investing.com}
\usage{
getEarnings(
  from = Sys.Date() - 5,
  to = Sys.Date() + 5,
  country = "United States"
)
}
\arguments{
\item{from}{the start date in yyyy-mm-dd format (default is 10 days prior to current date)}

\item{to}{the end date in yyyy-mm-dd format (default is the current date)}

\item{country}{the country from which to get earnings data (default is United States). One from Argentina,Australia,Austria,Bahrain,Belgium,Bosnia-Herzegovina,Botswana,Brazil,Bulgaria,Canada,Chile,China,Colombia,Costa Rica,Croatia,Cyprus,Czech Republic,Denmark,Egypt,Finland,France,Germany,Greece,Hong Kong,Hungary,Iceland,India,Indonesia,Ireland,Israel,Italy,Japan,Jordan,Kenya,Kuwait,Lebanon,Luxembourg,Malaysia,Malta,Mauritius,Mexico,Morocco,Namibia,Netherlands,New Zealand,Nigeria,Norway,Oman,Pakistan,Palestinian Territory,Peru,Philippines,Poland,Portugal,Qatar,Romania,Russia,Saudi Arabia,Serbia,Singapore,Slovakia,Slovenia,South Africa,South Korea,Spain,Sri Lanka,Sweden,Switzerland,Taiwan,Thailand,Tunisia,Turkey,Uganda,Ukraine,United Arab Emirates,United Kingdom,United States,Venezuela,Vietnam,Zimbabwe}
}
\value{
a data frame with earnings data
}
\description{
This function retrieves earnings data from the investing.com website for a specified time period
}
\note{
Not for the faint of heart. All profits and losses related are yours and yours alone. If you don't like it, write it yourself.
}
\examples{
getEarnings(from = '2023-02-07',to = '2023-02-13',country='Belgium')
}
\author{
Vyacheslav Arbuzov
}
