% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIPO.R
\name{getIPO}
\alias{getIPO}
\title{Get IPO calendar data from Investing.com}
\usage{
getIPO(from = Sys.Date() - 10, to = Sys.Date())
}
\arguments{
\item{from}{The start date of the date range in YYYY-MM-DD format. Default is the current date minus 10 days.}

\item{to}{The end date of the date range in YYYY-MM-DD format. Default is the current date.}
}
\value{
A data frame containing IPO calendar data for the specified date range.
}
\description{
This function retrieves IPO calendar data from Investing.com for a specified date range
}
\note{
Not for the faint of heart. All profits and losses related are yours and yours alone. If you don't like it, write it yourself.
}
\examples{
getIPO(from=Sys.Date(),to=Sys.Date()+3)
}
\author{
Vyacheslav Arbuzov
}
