\name{getOrders}
\alias{getOrders}
\title{Download order information from account}
\description{
  This function responsible for downloading orders information from account
}
\usage{
getOrders(Connection)
}
\arguments{
  \item{Connection}{localhost for terminal connections and name of exchange for web exchange connections}
 }
\keyword{getOrders}