% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSplits.R
\name{getSplits}
\alias{getSplits}
\title{Get stock split calendar data from investing.com}
\usage{
getSplits(from = Sys.Date() - 10, to = Sys.Date())
}
\arguments{
\item{from}{A date in the format 'YYYY-MM-DD' representing the start of the date range to retrieve data for (default is Sys.Date()-10).}

\item{to}{A date in the format 'YYYY-MM-DD' representing the end of the date range to retrieve data for (default is Sys.Date()).}
}
\value{
A data.table object containing the stock split calendar data from investing.com.
}
\description{
This function retrieves the stock split calendar data from investing.com between two given dates.
}
\note{
Not for the faint of heart. All profits and losses related are yours and yours alone. If you don't like it, write it yourself.
}
\examples{
getSplits(from=Sys.Date(),to=Sys.Date()+3)
}
\author{
Vyacheslav Arbuzov
}
