\name{getTradelog}
\alias{getTradelog}
\title{Load trade data}
\description{
  Function to load and manage trade data.
}
\usage{
getTradelog(Symbols = NULL,  from=Sys.Date(), to = Sys.Date())
}
\arguments{
  \item{Symbols}{a character vector specifying the names of each symbol (ticker) to be loaded.}
  \item{from}{start date of analyzing period.}
  \item{to}{end date of analyzing period.}
 }
\keyword{getTradelog}