\name{loadStockListMoex}
\alias{loadStockListMoex}
\title{Load stock list from MOEX}
\description{
Load stock list from MOEX. 
}
\usage{
loadStockListMoex(verbose = FALSE, market=c("MOEX", "FORTS"))
}
\arguments{
\item{market}{ load STOCK  and/or FORTS data}
}
\examples{
loadStockListMoex(verbose = FALSE, market=c("MOEX", "FORTS"))
}
\keyword{loadStockListMoex}