\name{openOrder}
\alias{openOrder}
\title{Create connection to exchange}
\description{
  This function responsible for creation connections.
}
\usage{
openOrder(Connection, Order)
}
\arguments{
  \item{Connection}{localhost for terminal connections and name of exchange for web exchange connections}
  \item{Order}{object type Order}
 }
\keyword{openOrder}