\name{Order}
\alias{Order}
\title{Create object type order}
\description{
  This function responsible for creation orders.
}
\usage{
Order(Symbol = "",Action = "BUY",Quantity = 10,OrderType = "LMT",Price = 0)
}
\arguments{
  \item{Symbol}{Symbol}
  \item{Action}{Action}
  \item{Quantity}{Quantity}
  \item{OrderType}{OrderType}
  \item{Price}{Price}
 }
\keyword{Order}




