% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rusquant-package.R
\docType{package}
\name{rusquant-package}
\alias{rusquant-package}
\alias{rusquant}
\title{Financial Modelling Framework extension for Russian stock market}
\description{
This is an extension for quantmod framework to simplify working with Russian
stock market.
}
\details{
\tabular{ll}{ Package: \tab rusquant\cr Type: \tab Package\cr Version: \tab
1.0\cr Date: \tab 2011-08-07\cr License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr } Please install it along with quantmod package. To be
done.
}
\references{
http://algorithmist.ru
}
\author{
Sergey Edunov <edunov@gmail.com> Vyacheslav Arbuzov
<arbuzov1989@gmail.com>
}
\keyword{market}
\keyword{package}
\keyword{russian}
\keyword{stock}
