\name{searchSymbol}
\alias{searchSymbol}
\title{Search Symbol name within MFD symbol list}
\description{
Search and return  part or full name of Symbol
}
\usage{
searchSymbol(Symbol)
}
\arguments{
\item{Symbol}{full or part Symbol name}
}
\examples{
searchSymbol("GAZP")
}
\keyword{searchSymbol}