% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickers.R
\docType{data}
\name{tickers}
\alias{tickers}
\title{Available tickers for provided datasources.}
\format{A data frame with 10157 rows and 4 variables:
\describe{
  \item{Group}{...}
  \item{Unknown}{...}
  \item{Id}{Ticker id}
  \item{Ticker}{Ticker name}
}}
\source{
...
}
\usage{
data(tickers)
}
\description{
Available tickers for provided datasources.
}
\keyword{datasets}
