% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth}
\alias{auth}
\title{Set creds for Datasource}
\usage{
auth(src = "Moex", login, password)
}
\arguments{
\item{src}{datasource name.}

\item{login}{login of datasource}

\item{password}{password of datasource}
}
\value{
Invisible NULL, side-effect function setting an environment variable.
}
\description{
Sets creds for the Datasource  by storing it in an
environment variable for the current R session. This token will be used by
other functions in the package to authenticate API requests.
}
\examples{
\dontrun{
  auth(login = "user@email.com",password = "mypassword")
}
}
