\name{cancelOrder}
\alias{cancelOrder}
\title{Close order}
\description{
  This function responsible for cancel order
}
\usage{
cancelOrder(Connection, OrderID)
}
\arguments{
  \item{Connection}{localhost for terminal connections and name of exchange for web exchange connections}
  \item{OrderID}{order id}
 }
\keyword{cancelOrder}