% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEarnings.R
\name{getEarnings}
\alias{getEarnings}
\title{Download earnings data from investing.com}
\usage{
getEarnings(
  from = Sys.Date() - 5,
  to = Sys.Date() + 5,
  country = "United States"
)
}
\arguments{
\item{from}{the start date in yyyy-mm-dd format (default is 10 days prior to current date)}

\item{to}{the end date in yyyy-mm-dd format (default is the current date)}

\item{country}{the country from which to get earnings data (default is United States)}
}
\value{
a data frame with earnings data
}
\description{
This function retrieves earnings data from the investing.com website for a specified time period
}
\note{
Not for the faint of heart. All profits and losses related are yours and yours alone. If you don't like it, write it yourself.
}
\examples{
getEarnings(from = Sys.Date(),to = Sys.Date()+3,country='Belgium')
}
\author{
Vyacheslav Arbuzov
}
