% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSymbols.Algopack.R
\name{getSymbols.Algopack}
\alias{getSymbols.Algopack}
\title{Download AlgoPack data from MOEX}
\usage{
getSymbols.Algopack(
  Symbols,
  env = globalenv(),
  from = Sys.Date() - 30,
  to = Sys.Date(),
  verbose = TRUE,
  type = "tradestats",
  login = "user@email.com",
  password = "mypassword",
  auto.assign = FALSE,
  ...
)
}
\arguments{
\item{Symbols}{a character vector of AlgoPack symbols to download data for.}

\item{env}{environment where to create the downloaded data object.}

\item{from}{a character string indicating the start date of the data to download, in YYYY-MM-DD format.}

\item{to}{a character string indicating the end date of the data to download, in YYYY-MM-DD format.}

\item{verbose}{a logical indicating whether to print the response details or not.}

\item{type}{a character string indicating the AlgoPack type possible values are c('tradestats','orderstats','obstats'), default 'tradestats'.}

\item{login}{a character string with login from ISS Moex}

\item{password}{a character string with password from ISS Moex}

\item{auto.assign}{a logical indicating whether to automatically assign the downloaded data to the global environment.}

\item{...}{additional arguments passed to getSymbols.AlgoPack}
}
\value{
returns an data.table object containing financial data
}
\description{
Download historical market data from AlgoPack  for a given symbol and time range.
}
\note{
Not for the faint of heart. All profits and losses related are yours and yours alone. If you don't like it, write it yourself.
}
\examples{
getSymbols.Algopack('SBER',from = '2023-10-24',to='2023-11-04',login = 'my-email.com',password = 'mypassword')
getSymbols('SBER',src='Algopack',login = 'my-email.com',password = 'mypassword')
}
\author{
Vyacheslav Arbuzov
}
