% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSymbols.R
\name{getSymbols}
\alias{getSymbols}
\alias{getSymbols.Poloniex}
\alias{getSymbols.Alor}
\alias{getSymbols.rogov}
\alias{getSymbols.Finam}
\alias{getSymbols.mfd}
\alias{getSymbols.Forts}
\title{Load and manage market data}
\usage{
getSymbols(Symbols = NULL, from='2007-01-01', to = Sys.Date(),
                  src = "yahoo", period = "day")

getSymbols.Poloniex(Symbols, env, return.class = "xts",
  index.class = "Date", from = "2007-01-01", to = Sys.Date(),
  adjust = FALSE, period = "day", ...)

getSymbols.Alor(Symbols, env, return.class = "xts", index.class = "Date",
  from = "2007-01-01", to = Sys.Date(), adjust = FALSE, period = "day",
  ...)

getSymbols.rogov(Symbols, env, return.class = "xts", index.class = "Date",
  from = "2007-01-01", to = Sys.Date(), adjust = FALSE, period = "day",
  ...)

getSymbols.Finam(Symbols, env, return.class = "xts", index.class = "Date",
  from = "2007-01-01", to = Sys.Date(), adjust = FALSE, period = "day",
  ...)

getSymbols.mfd(Symbols, env, return.class = "xts", index.class = "Date",
  from = "2007-01-01", to = Sys.Date(), adjust = FALSE, period = "day",
  ...)

getSymbols.Forts(Symbols, env, return.class = "xts", index.class = "Date",
  from = "2007-01-01", to = Sys.Date(), adjust = FALSE, period = "day",
  ...)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol
(ticker) to be loaded.}

\item{from}{start date of analyzing period.}

\item{to}{end date of analyzing period.}

\item{period}{calendar periods for Finam ("tick", "secs", "seconds", "mins",
"hours", "days", "weeks", "months", "quarters", "years").}

\item{src}{character string specifying sourcing method (only yahoo, Finam,
mfd, rogov, oanda, google).}
}
\description{
Function to load and manage data. Current src methods available are: yahoo,
Finam, mfd, rogov, oanda, google. Current period methods available are:
sec, hour, day, week, month, year.
}
\examples{
getSymbols('AFLT', src='Finam')
}
