\name{openOrder}
\alias{openOrder}
\title{Send order to exchange}
\description{
  This function responsible for sending orders to exchanges.
}
\usage{
openOrder(Connection, Order)
}
\arguments{
  \item{Connection}{object ConnectionToExchange by function to Connect}
  \item{Order}{object type Order}
 }
\keyword{openOrder}